Add ActiveArea to image struct
At least for dng files the lens correction data seem to be not defined for
default crop but for the active area.
So we read that tag and store it in the image struct, the rawspeed reads
always the DefaultCrops and leaves it to the user in the hidden slider section.

Will allow the rawprepare module to use approprioate crops depending on
lens correction mode at least, could also bump the pawprepare module and
do some overriding.