Picker and blending support for some rgb profiles
See #19724 for .icc file and issue description.

dt expects a "working matrix" when used for pickers or blending, possibly .icc files don't provide them leading to
issues in modules before colorin in the pipe,
- module pickers and blending does not work for IOP_CS_JZCZHZ.

If a rgb profile file is added to the list we check for missing matrixes, if that is the case we insert a "pass matrix"
and report that in the log.

Some maintainance and improved logs while being here.