Improved debugging support for colorin profile
We might set a profile in colorin that does not include in&out matrixes.

This is not relevant in most cases as we do a fallback to lcms2 when changing the module's
colorspace.

But it becomes relevant for blending and picker reading if the module is before colorin in the
pipe. Both mentioned algos currently rely on a matrix and thus fail is sucha profile has been selected.

At least we should provide information via the log and control_log so the user knows why that fails
and we can further investigate,