Avoid control threads race condition
When starting the dt gui and calling `gtk_main()` we must be sure we can safely dispatch a control job.
We can do so only if all control threads are running.

We achieve this by adding a control->running atomic counter which is increased whenever any of the threads
code has been started. As we know the number of started threads we can check & wait (for up to 1sec) until
all threads are up&running before leaving `dt_control_jobs_init()`.

The new function `gboolean dt_control_all_running()` tests for running vs requested threads and is
used to test if we can finally use gtk_main().